+++ filter_test.go
@@ -200,43 +200,3 @@ func TestFiltersRemoveRules(t *testing.T) {
 		}
 	}
 }
-
-func TestMultiaddrToIPNet(t *testing.T) {
-	type testCase struct {
-		name      string
-		ma        string
-		ips       []string
-		contained []bool
-	}
-
-	testCases := []testCase{
-		{
-			ma:        "/ip4/1.2.3.0/ipcidr/24",
-			ips:       []string{"1.2.3.4", "1.2.3.9", "2.1.1.1"},
-			contained: []bool{true, true, false},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			ma, err := NewMultiaddr(tc.ma)
-			if err != nil {
-				t.Fatalf("failed to parse multiaddr %v", ma)
-			}
-
-			ipnet, err := MultiaddrToIPNet(ma)
-			if err != nil {
-				t.Fatalf("failed to parse multiaddr %v into ipnet", ma)
-			}
-			for i, ipString := range tc.ips {
-				ip := net.ParseIP(ipString)
-				if ip == nil {
-					t.Fatalf("failed to parse IP %s", ipString)
-				}
-				if ipnet.Contains(ip) != tc.contained[i] {
-					t.Fatalf("Contains check failed. Expected %v got %v", tc.contained[i], ipnet.Contains(ip))
-				}
-			}
-		})
-	}
-}
