+++ net/convert_test.go
@@ -5,6 +5,7 @@ import (
 	"runtime"
 	"testing"
 
+	"github.com/multiformats/go-multiaddr"
 	ma "github.com/multiformats/go-multiaddr"
 )
 
@@ -201,4 +202,41 @@ func TestDialArgs(t *testing.T) {
 	test("/dns4/abc.com", "ip4", "abc.com")                         // Just DNS4
 	test("/dns6/abc.com/udp/1234", "udp6", "abc.com:1234")          // DNS6:port
 	test("/dns6/abc.com", "ip6", "abc.com")                         // Just DNS6
+}
+
+func TestMultiaddrToIPNet(t *testing.T) {
+	type testCase struct {
+		name      string
+		ma        string
+		ips       []string
+		contained []bool
+	}
+
+	testCases := []testCase{
+		{
+			ma:        "/ip4/1.2.3.0/ipcidr/24",
+			ips:       []string{"1.2.3.4", "1.2.3.9", "2.1.1.1"},
+			contained: []bool{true, true, false},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			ma := multiaddr.StringCast(tc.ma)
+
+			ipnet, err := MultiaddrToIPNet(ma)
+			if err != nil {
+				t.Fatalf("failed to parse multiaddr %v into ipnet", ma)
+			}
+			for i, ipString := range tc.ips {
+				ip := net.ParseIP(ipString)
+				if ip == nil {
+					t.Fatalf("failed to parse IP %s", ipString)
+				}
+				if ipnet.Contains(ip) != tc.contained[i] {
+					t.Fatalf("Contains check failed. Expected %v got %v", tc.contained[i], ipnet.Contains(ip))
+				}
+			}
+		})
+	}
 }
